% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_diff.R
\name{ci_diff}
\alias{ci_diff}
\title{Confidence interval of AUC difference using unbiased variance estimator}
\usage{
ci_diff(ROC1,ROC2,type=c("two.sided","less","greater"),level=NULL)
}
\arguments{
\item{ROC1, ROC2}{Two \code{ROC} objects referring two ROC curves.}

\item{type}{Character. Type of the confidence interval. Must be one of 'two.sided', 'less', 'greater'. 'less',
'greater' refer to one-sided confidence interval corresponding to the alternatives. The first letter is
sufficient.}

\item{level}{A numeric scalar in (0, 1). Indicating the level of confidence interval. Default is 0.95.}
}
\value{
A vector of length 2 including the lower bound and upper bound of the confidence interval.
}
\description{
Calculate the confidence interval of the AUC difference using normal approximation with the unbiased variance estimator
}
\section{Warning}{

When \eqn{AUC_1=AUC_2}, the variance of the AUC difference may be 0, so the confidence interval should always
have width 0 and can be misleading.
}

\examples{
library(uROC)

set.seed(321)
## Generate data
Data=SimuCaseCont(30,30,0.9,0.9,0.87,0.90,"c","t",marg1="Normal",marg2="Normal")
## Construct ROC curves
uroc1=ROC(Data[,3],Data[,1],case_score_ind = "higher")
uroc2=ROC(Data[,3],Data[,2],case_score_ind = "higher")
## Two-sided confidence interval
ci_diff(uroc1,uroc2,type = "two.sided")

## One-sided confidence interval
ci_diff(uroc1,uroc2,type="less",level=0.95)
}
\references{
Lu, Y. and Shao, Y. (2020). ucompROC: A new powerful test to compare correlated ROC curves.
}
\seealso{
\code{\link{ucompROC}}
}
