% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uroc.R
\name{plot.uroc}
\alias{plot.uroc}
\title{Plot of UROC curve.}
\usage{
\method{plot}{uroc}(x, ...)
}
\arguments{
\item{x}{object of class "uroc"}

\item{...}{further arguments to \code{\link{plot}}}
}
\value{
plot
}
\description{
This function plots a UROC curve.
}
