% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocm.R
\name{rocm}
\alias{rocm}
\title{Builds the sequence of ROC curves for the ROC movie (ROCM)}
\usage{
rocm(response, predictor, b = 100, a = 400)
}
\arguments{
\item{response}{a numeric vector of real valued responses}

\item{predictor}{a numeric vector of the same length than \code{response}, containing real valued predictions for each observation}

\item{b}{default is \code{b=100} Used to select a subset of all ROC curves for the ROC movie with at least \code{a} and at most \code{a+b} ROC curves}

\item{a}{default is \code{a=400}. Used to select a subset of all ROC curves for the ROC movie with at least \code{a} and at most \code{a+b} ROC curves}
}
\value{
List
}
\description{
This function computes a List of ROC curves.
}
\details{
The ROC movie can be used to visualize the performance of a real valued foreacsting problem. Therefore, a sequence of ROC curves is generated which should than be combined into a GIF animation. Each entry of the list consist of two vectors of length 1000 containing the values of farate (1-Specificity) and hitrate (sensitivity) and three values, namely the associated auc value, the weight and the threshold
}
\examples{
\dontrun{
data(longley)
response = longley$Employed
predictor = longley$GNP
rocm(response, predictor)}
}
