% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimuCaseCont.R
\name{SimuCaseCont}
\alias{SimuCaseCont}
\title{Random generation of two correlated risk scores under case-control settings}
\usage{
SimuCaseCont(m,n,tau1,tau0,AUC_1,AUC_2,
  cop1=c("g","t","c"),cop0=c("g","t","c"),marg1=c("Normal","Exp"),marg2=c("Normal","Exp"))
}
\arguments{
\item{m, n}{Integers. Specify the number of cases(m), number of controls(n).}

\item{tau1, tau0}{Numeric. The Kendall's tau correlation between the two risk scores in cases(tau1)
and controls(tau0) separately. Should be specified in (0, 1).}

\item{AUC_1, AUC_2}{The AUC of each of the risk scores. Should be specified (0.5, 1)}

\item{cop1, cop0}{Character. Specify the copula for cases(cop1) and controls(cop0). Choosing from
'g'(Gaussian copula), 't'(t copula), and 'c'(Clayton copula).}

\item{marg1, marg2}{Character. Marginal distributions for two risk scores, respectively. Both choose from "Normal"
or "Exp"(exponential).}
}
\value{
A \eqn{(m+n) \times 3} data matrix. The first two columns are the generated values for risk score 1
and 2, accordingly, and the last column is the binary response.
}
\description{
Generate two correlated risk score with given AUCs and correlations under case-control setting using copula theory.
}
\examples{
library(uROC)

set.seed(123)
## Generate data: m=n=100, tau1=0.8, tau0=0.9, AUC_1=AUC_2=0.9, Clayton copula in cases,
## t copula in controls, both normal margins for score 1 and 2
Data=SimuCaseCont(100,100,0.8,0.9,0.9,0.9,"c","t",marg1="Normal",marg2="Normal")
Data
## Check Kendall's tau correlation
cor(Data[1:100,1],Data[1:100,2],method="kendall")
cor(Data[101:200,1],Data[101:200,2],method="kendall")
## Check AUC
roc1=ROC(Data[,3],Data[,1],case_score_ind = "higher")
roc2=ROC(Data[,3],Data[,2],case_score_ind = "higher")
roc1$auc
roc2$auc
}
\references{
Nagler T., Schepsmeier U., Stoeber J. , Brechmann E C., Graeler B. and Erhardt T. (2019).
\emph{VineCopula: Statistical Inference of Vine Copulas.} R package version 2.3.0.
https://CRAN.R-project.org/package=VineCopula
}
