% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucov.R
\name{ucov}
\alias{ucov}
\title{Unbiased covariance of AUCs under two ROC curves}
\usage{
ucov(ROC1, ROC2)
}
\arguments{
\item{ROC1, ROC2}{Two \code{ROC} objects referring two ROC curves.}
}
\value{
A numeric value.
}
\description{
Calculate unbiased covariance of AUCs under two ROC curves.
}
\section{Warning}{

Covariance of two AUCs with any AUC=1 is always 0 and can be misleading.
}

\examples{
set.seed(123)
## Generate data:
Data=SimuCaseCont(100,100,0.8,0.9,0.9,0.9,"c","t",marg1="Normal",marg2="Normal")
## Construct ROC curves
roc1=ROC(Data[,3],Data[,1],case_score_ind = "higher")
roc2=ROC(Data[,3],Data[,2],case_score_ind = "higher",ci=TRUE,level=0.95)
## Compute the covariance of two ROC curves
ucov(roc1,roc2)
}
\references{
Lu, Y. and Shao, Y. (2020). ucompROC: A new powerful test to compare correlated ROC curves.
}
