% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uroc.R
\name{uroc}
\alias{uroc}
\title{Computes a UROC curve.}
\usage{
uroc(response, predictor, approx = FALSE, split = 1)
}
\arguments{
\item{response}{a numeric vector of real valued responses.}

\item{predictor}{a numeric vector of the same length than \code{response}, containing real valued predictions for each observation.}

\item{approx}{Boolean. If TRUE approximates true roc curve with faster algorithm.}

\item{split}{integer value with a default of \code{split = 1}. Computes uroc curve by considering only a subset of all N-1 available ROC curves to reduce computation time. The split parameter defines the distance between a set of equidistant indices which are then used to select particular ROC curves among the N-1.}
}
\value{
If \code{object = TRUE} this function returns a list of class "uroc".
}
\description{
This function builds a UROC curve and returns a "uroc" object, a list of class "uroc".
}
\details{
The default option to compute uroc curve generates an approximation to the UROC curve by using linear interpolation of each ROC curve. For small datasets or binary response the exact uroc curve is computed. Setting option \code{approx = TRUE} uses a faster approximation algorithm where computation time can be further reduced by setting the paramter \code{split} which selects only a subset of ROC curves in the computation.
}
\examples{
data(longley)
response = longley$Employed
predictor = longley$GNP
uroc(response, predictor)
}
