% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvar.R
\name{uvar}
\alias{uvar}
\title{Unbiased variance of a single ROC curve.}
\usage{
uvar(ROC)
}
\arguments{
\item{ROC}{An \code{ROC} object obtained by ROC.}
}
\value{
A numeric value.
}
\description{
Calculate unbiased variance of a single ROC curve.
}
\section{Warning}{

Variance of AUC=1 is always 0 and can be misleading.
}

\examples{
set.seed(123)
## Generate data:
Data=SimuCaseCont(100,100,0.8,0.9,0.9,0.9,"c","t",marg1="Normal",marg2="Normal")
## Construct ROC curves
roc1=ROC(Data[,3],Data[,1],case_score_ind = "higher")
roc2=ROC(Data[,3],Data[,2],case_score_ind = "higher",ci=TRUE,level=0.95)
## Calculate the unbiased variance of two ROC curves
uvar(roc1)
uvar(roc2)
}
\references{
Lu, Y. and Shao, Y. (2020). Preprint
}
