% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpa.R
\name{cpa}
\alias{cpa}
\alias{cpa.default}
\alias{cpa.uroc}
\title{Computes coefficient of predictive ability (CPA)}
\usage{
cpa(...)

\method{cpa}{default}(response, predictor, ...)

\method{cpa}{uroc}(uroc, ...)
}
\arguments{
\item{...}{ignored}

\item{response}{a numeric vector of real valued responses}

\item{predictor}{a numeric vector of the same length as \code{response}, containing real valued predictions for each observation}

\item{uroc}{an object of class "uroc" contaning the values of the false alarm rate (1-specificity) and the hitrate (sensitivity) of the UROC curve}
}
\value{
The numeric CPA value
}
\description{
This function computes the coefficient of predictive ability which is equalivalent to the area under the UROC curve. Two syntaxes are possible: one object of class "uroc" or two vectors, the response and the predictor.
}
\details{
The CPA is an asymmetric measure that is linearly related to the correlation between the classes of the response variable and the ranks of the predictor
}
\examples{

data(longley)
response = longley$Employed
predictor = longley$GNP
cpa(response, predictor)
}
