% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocm.R
\name{rocm}
\alias{rocm}
\title{Builds a ROC movie (ROCM)}
\usage{
rocm(response, predictor, path, output = "animation.gif", clean = TRUE,
  convert = "convert", cmd.fun = if (.Platform$OS.type == "windows")
  shell else system, interval = 0.1, ...)
}
\arguments{
\item{response}{a numeric vector of real valued responses}

\item{predictor}{a numeric vector of the same length than \code{response}, containing real valued predictions for each observation}

\item{path}{folder path}

\item{output}{name of GIF animation}

\item{clean}{if FALSE png files are not deleted}

\item{convert}{convert command for the function  \link[animation]{im.convert}}

\item{cmd.fun}{a function to invoke OS command in  \link[animation]{im.convert}}

\item{interval}{a postve number to set the time interval of the animation (unit in second) in \link[animation]{ani.options}}

\item{...}{plotting arguments}
}
\value{
GIF animation
}
\description{
This function computes an animated ROC movie.
}
\details{
The ROC movie can be used to visualize the performance of a real valued foreacsting problem. Therefore, a sequence of png-files is generated and combined into a GIF animation using the external software ImageMagick (\url{https://imagemagick.org/}) and the R package \code{animation}.
}
\examples{
\dontrun{
data(longley)
response = longley$Employed
predictor = longley$GNP
rocm(response, predictor, path="/home")}
}
