% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uroc.R
\name{uroc}
\alias{uroc}
\title{Computes a UROC curve}
\usage{
uroc(response, predictor, object = FALSE, plot = TRUE,
  algo = "approx", split = 1)
}
\arguments{
\item{response}{a numeric vector of real valued responses}

\item{predictor}{a numeric vector of the same length than \code{response}, containing real valued predictions for each observation}

\item{object}{if TRUE an object of type uroc is returned containg the false alarm rate and the hitrate of the UROC curve}

\item{plot}{plot the UROC curve? if \code{FALSE} the curve is not displayed}

\item{algo}{optional argument to select an algorithm for the computation of the UROC curve. See Details.}

\item{split}{a integer value with a default of \code{split = 1}. Computes uroc curve by considering only a subset of all N-1 available ROC curves to reduce computation time. The split parameter defines the distance between a set of equidistant indices which are then used to select particular ROC curves among the N-1.}
}
\value{
If \code{object = TRUE} this function returns a list of class "uroc".
}
\description{
This function builds a UROC curve and returns a "uroc" object, a list of class "uroc".
}
\details{
There are 2 different algorithms available to create a uROC curve. The default option is  \code{algo="approx"} which generates an approximation to the UROC curve by using linear interpolation of each ROC curve. To reduce computation time the paramter \code{split} can be specified to select a subset of ROC curves in the computation. The input argument \code{algo="exact"} computes the exact UROC curve and should only be used on small data.
}
\examples{
data(longley)
response = longley$Employed
predictor = longley$GNP
uroc(response, predictor)
}
